Internal: rename `Curves` to `CurvesWithOID`
Following the addition of the new format for Montgomery curves,
which do not rely on OIDs.
